wip(#1352): add participants and voting handler
